requirements.txt
Django==1.7.7
Pillow==2.7.0
django-registration-redux==1.1
django-geoposition==0.2.2
